split program factories
