+++ ttnn/cpp/ttnn/operations/normalization/layernorm_distributed/device/layernorm_post_all_gather_welford_program_factory.hpp
@@ -5,21 +5,20 @@
 #pragma once
 
 #include "layernorm_post_all_gather_device_operation_types.hpp"
-#include "layernorm_post_all_gather_welford_program_factory.hpp"
 #include "ttnn/device_operation.hpp"
 
 namespace ttnn::operations::normalization::program {
 
-// Shared variables for normal program factory
-struct LayerNormPostAllGatherSharedVariables {
+// Shared variables for Welford program factory
+struct LayerNormPostAllGatherWelfordSharedVariables {
     tt::tt_metal::KernelHandle reader_kernel_id = 0;
     tt::tt_metal::KernelHandle writer_kernel_id = 0;
     std::vector<CoreCoord> cores;
 };
 
-// Program factory for normal (non-Welford) operation
-struct LayerNormPostAllGatherProgramFactory {
-    using shared_variables_t = LayerNormPostAllGatherSharedVariables;
+// Program factory for Welford algorithm (layernorm only)
+struct LayerNormPostAllGatherWelfordProgramFactory {
+    using shared_variables_t = LayerNormPostAllGatherWelfordSharedVariables;
     using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
 
     static cached_program_t create(
